requirements.txt
argparse==1.4.1; python_version < '2.7'
beautifulsoup4==4.6.0
keyring==8.4.1
ofxhome==0.3.3
ofxparse==0.14
lxml>=3.5.0
keyrings.alt==1.1.1
pycrypto==2.6.1
psycopg2==2.7.4
pyyaml=3.12